requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-genai==1.10.0
yt-dlp==2024.11.18
pydantic==2.9.2
python-multipart==0.0.12
python-dotenv==1.0.1